[#601] tests: transfer unittests to dependant apps
