+++ cove_360/tests.py
@@ -1,0 +1,74 @@
+import os
+
+import pytest
+from django.core.files.base import ContentFile
+from django.core.files.uploadedfile import UploadedFile
+
+from . import settings
+from . lib.schema import Schema360
+from cove.input.models import SuppliedData
+
+
+@pytest.mark.django_db
+def test_explore_page(client):
+    data = SuppliedData.objects.create()
+    data.original_file.save('test.json', ContentFile('{}'))
+    data.current_app = 'cove_360'
+    resp = client.get(data.get_absolute_url())
+    assert resp.status_code == 200
+    assert resp.context['conversion'] == 'flattenable'
+
+    resp = client.post(data.get_absolute_url(), {'flatten': 'true'})
+    assert resp.status_code == 200
+    assert resp.context['conversion'] == 'flatten'
+    assert 'converted_file_size' in resp.context
+    assert 'converted_file_size_titles' in resp.context
+
+
+@pytest.mark.django_db
+def test_explore_page_csv(client):
+    data = SuppliedData.objects.create()
+    data.original_file.save('test.csv', ContentFile('a,b'))
+    resp = client.get(data.get_absolute_url())
+    assert resp.status_code == 200
+    assert resp.context['conversion'] == 'unflatten'
+    assert resp.context['converted_file_size'] == 20
+
+
+@pytest.mark.django_db
+def test_explore_not_json(client):
+    data = SuppliedData.objects.create()
+    with open(os.path.join('cove_360', 'fixtures', 'WellcomeTrust-grants_malformed.json')) as fp:
+        data.original_file.save('test.json', UploadedFile(fp))
+    resp = client.get(data.get_absolute_url())
+    assert resp.status_code == 200
+    assert b'not well formed JSON' in resp.content
+
+
+@pytest.mark.django_db
+def test_explore_unconvertable_spreadsheet(client):
+    data = SuppliedData.objects.create()
+    with open(os.path.join('cove', 'fixtures', 'bad.xlsx'), 'rb') as fp:
+        data.original_file.save('basic.xlsx', UploadedFile(fp))
+    resp = client.get(data.get_absolute_url())
+    assert resp.status_code == 200
+    assert b'We think you tried to supply a spreadsheet, but we failed to convert it to JSON.' in resp.content
+
+
+@pytest.mark.django_db
+def test_explore_unconvertable_json(client):
+    data = SuppliedData.objects.create()
+    with open(os.path.join('cove', 'fixtures', 'unconvertable_json.json')) as fp:
+        data.original_file.save('unconvertable_json.json', UploadedFile(fp))
+    resp = client.post(data.get_absolute_url(), {'flatten': 'true'})
+    assert resp.status_code == 200
+    assert b'could not be converted' in resp.content
+
+
+def test_schema_360():
+    schema = Schema360()
+    assert schema.release_schema_name == settings.COVE_CONFIG['schema_item_name']
+    assert schema.release_pkg_schema_name == settings.COVE_CONFIG['schema_name']
+    assert schema.schema_host == settings.COVE_CONFIG['schema_host']
+    assert schema.release_schema_url == settings.COVE_CONFIG['schema_host'] + settings.COVE_CONFIG['schema_item_name']
+    assert schema.release_pkg_schema_url == settings.COVE_CONFIG['schema_host'] + settings.COVE_CONFIG['schema_name']
