+++ cove_ocds/views.py
@@ -112,7 +112,7 @@ def explore_ocds(request, pk):
                         schema_ocds.create_extended_release_schema_file(db_data.upload_dir(), db_data.upload_url())
                         url = schema_ocds.extended_schema_file
 
-                context.update(convert_json(request, db_data, url, replace=replace))
+                context.update(convert_json(request, db_data, schema_url=url, replace=replace))
 
     else:
         select_version = post_version_choice or db_data.schema_version
@@ -120,7 +120,7 @@ def explore_ocds(request, pk):
         # Replace json conversion when user chooses a different schema version.
         if db_data.schema_version and schema_ocds.version != db_data.schema_version:
             replace = True
-        context.update(convert_spreadsheet(request, db_data, file_type, schema_ocds.release_schema_url, replace=replace))
+        context.update(convert_spreadsheet(request, db_data, file_type, schema_url=schema_ocds.release_schema_url, replace=replace))
         with open(context['converted_path'], encoding='utf-8') as fp:
             json_data = json.load(fp)
